requirements.txt
flask
psutil
pandas
matplotlib
pywebview